Use a Swift package instead of an xcode project
