--- raytracer/Background.swift
@@ -1,19 +1,0 @@
-import Foundation
-
-protocol Background {
-  func colorFrom(_ ray: Ray) -> Color
-}
-
-struct Sky: Background {
-  let top, bottom: Color
-  
-  func colorFrom(_ ray: Ray) -> Color {
-    let t = 0.5 * (ray.direction.normalized().y + 1)
-    let color = lerpColor(bottom, top, t)
-    return Color(
-      r: ray.color.r * color.r,
-      g: ray.color.g * color.g,
-      b: ray.color.b * color.b
-    )
-  }
-}
